refactor(api): split Program.cs endpoints and enforce thin-entrypoint governance
- implement T070: extract auth/session endpoints to Auth/AuthEndpoints.cs
- implement T071: extract catalog and pedidos endpoints to dedicated endpoint modules
- reduce Program.cs from monolithic mapping to orchestration-only bootstrap
- update spec kit governance/templates with modularity requirements
- update feature plan/tasks with decomposition rules and mark T070/T071 done
