+++ src/backend/Versatus.ForcaVendas.Api/Program.cs
@@ -1,10 +1,9 @@
-using Versatus.ForcaVendas.Domain.Pedidos;
-using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
-using Microsoft.EntityFrameworkCore;
 using MediatR;
 using FluentValidation;
+using Microsoft.EntityFrameworkCore;
 using Prometheus;
+using Versatus.ForcaVendas.Api;
 using Versatus.ForcaVendas.Api.Health;
 using Versatus.ForcaVendas.Application.Catalogo;
 using Versatus.ForcaVendas.Application.Licenca;
@@ -132,565 +131,9 @@ app.MapGet("/health/ready", async (HealthCheckService hc) =>
 
 // Expose Prometheus metrics at /metrics
 app.MapMetrics();
-app.MapPost("/auth/login", async (
-    LoginRequest request,
-    IOptions<AuthOptions> options,
-    IJwtTokenService tokenService,
-    IRefreshTokenStore refreshTokenStore,
-    Versatus.ForcaVendas.Domain.Auth.IUsuarioRepository usuarioRepository,
-    ITenantSubscriptionRepository subscriptionRepository,
-    ISessionStore sessionStore,
-    ISessionAuditEventRepository auditRepo,
-    HttpContext httpContext,
-    CancellationToken cancellationToken) =>
-{
-    var errors = request.Validate();
-    if (errors.Count > 0)
-    {
-        return Results.ValidationProblem(errors);
-    }
-
-    var authOptions = options.Value;
-    if (authOptions.Tenants.Count == 0)
-    {
-        return Results.Problem(
-            detail: "Configuracao de autenticacao invalida no servidor.",
-            statusCode: StatusCodes.Status500InternalServerError);
-    }
-
-    var usuarioBanco = await usuarioRepository.GetByEmailAsync(request.Email, cancellationToken);
-    if (usuarioBanco is null)
-    {
-         return Results.Unauthorized();
-    }
-
-    if (!BCrypt.Net.BCrypt.Verify(request.Password, usuarioBanco.PasswordHash))
-    {
-         return Results.Unauthorized();
-    }
-
-    var user = new DemoUser
-    {
-        UserId = usuarioBanco.Id.ToString(),
-        TenantId = usuarioBanco.TenantId.ToString(),
-        Username = usuarioBanco.Username,
-        Email = usuarioBanco.Email,
-        Role = usuarioBanco.Role,
-        Password = ""
-    };
-
-    var tenantExists = authOptions.Tenants
-        .Any(t => string.Equals(t, user.TenantId, StringComparison.OrdinalIgnoreCase));
-
-    if (!tenantExists)
-    {
-        return Results.Unauthorized();
-    }
-
-    var subscription = await subscriptionRepository.GetByTenantIdAsync(user.TenantId, cancellationToken);
-    if (subscription is null || !subscription.IsActive)
-    {
-        return Results.StatusCode(StatusCodes.Status403Forbidden);
-    }
-
-    var activeSessions = await sessionStore.CountActiveAsync(user.TenantId, cancellationToken);
-    if (!subscription.HasAvailableSeats(activeSessions))
-    {
-        return Results.Problem(
-            detail: "Limite de usuarios simultaneos atingido para este tenant.",
-            statusCode: StatusCodes.Status403Forbidden,
-            title: "Concurrent user limit reached");
-    }
-
-    var tokenPair = tokenService.Generate(user);
-    refreshTokenStore.Save(tokenPair.RefreshToken, user.UserId, user.TenantId, tokenPair.RefreshTokenExpiresAtUtc);
-
-    await sessionStore.AddAsync(tokenPair.SessionId, user.UserId, user.TenantId, cancellationToken);
-
-    // Audit login event
-    var auditEvent = new Versatus.ForcaVendas.Domain.Auditoria.SessionAuditEvent(
-        Id: Guid.NewGuid().ToString(),
-        UserId: user.UserId,
-        TenantId: user.TenantId,
-        EventType: Versatus.ForcaVendas.Domain.Auditoria.SessionAuditEventType.Login,
-        Timestamp: DateTimeOffset.UtcNow,
-        IpAddress: httpContext.Connection.RemoteIpAddress?.ToString(),
-        UserAgent: httpContext.Request.Headers["User-Agent"].ToString()
-    );
-    await auditRepo.AddAsync(auditEvent, cancellationToken);
-
-    return Results.Ok(new LoginResponse(
-        tokenPair.AccessToken,
-        tokenPair.RefreshToken,
-        tokenPair.AccessTokenExpiresInSeconds,
-        "Bearer"));
-})
-.WithName("Login")
-.WithOpenApi();
- 
-app.MapPost("/auth/refresh", async (
-    RefreshTokenRequest request,
-    IOptions<AuthOptions> options,
-    IJwtTokenService tokenService,
-    Versatus.ForcaVendas.Domain.Auth.IUsuarioRepository usuarioRepository,
-    IRefreshTokenStore refreshTokenStore,
-    CancellationToken cancellationToken) =>
-{
-    var errors = request.Validate();
-    if (errors.Count > 0)
-    {
-        return Results.ValidationProblem(errors);
-    }
-
-    if (!refreshTokenStore.TryGetActive(request.RefreshToken, out var tokenInfo))
-    {
-        return Results.Unauthorized();
-    }
-
-    if (!Guid.TryParse(tokenInfo.UserId, out var userId))
-    {
-        return Results.Unauthorized();
-    }
-
-    var usuarioBanco = await usuarioRepository.GetByIdAsync(userId, cancellationToken);
-
-    if (usuarioBanco is null || !string.Equals(usuarioBanco.TenantId.ToString(), tokenInfo.TenantId, StringComparison.OrdinalIgnoreCase))
-    {
-        return Results.Unauthorized();
-    }
-
-    var user = new DemoUser
-    {
-        UserId = usuarioBanco.Id.ToString(),
-        TenantId = usuarioBanco.TenantId.ToString(),
-        Username = usuarioBanco.Username,
-        Email = usuarioBanco.Email,
-        Role = usuarioBanco.Role,
-        Password = ""
-    };
-
-    var tokenPair = tokenService.Generate(user);
-
-    // Rotate refresh token to reduce replay window.
-    refreshTokenStore.Revoke(request.RefreshToken);
-    refreshTokenStore.Save(tokenPair.RefreshToken, user.UserId, user.TenantId, tokenPair.RefreshTokenExpiresAtUtc);
-
-    return Results.Ok(new LoginResponse(
-        tokenPair.AccessToken,
-        tokenPair.RefreshToken,
-        tokenPair.AccessTokenExpiresInSeconds,
-        "Bearer"));
-})
-.WithName("RefreshToken")
-.WithOpenApi();
-
-app.MapGet("/tenant/ping", (ITenantContext tenantContext) =>
-{
-    return Results.Ok(new
-    {
-        message = "Tenant context resolved.",
-        tenantId = tenantContext.TenantId
-    });
-})
-.WithName("TenantPing")
-.WithOpenApi();
-
-app.MapGet("/licenca/{tenantId}/limite", async (
-    string tenantId,
-    ITenantSubscriptionRepository repository,
-    CancellationToken cancellationToken) =>
-{
-    var subscription = await repository.GetByTenantIdAsync(tenantId, cancellationToken);
-    if (subscription is null)
-    {
-        return Results.NotFound(new { message = "Tenant nao encontrado." });
-    }
-
-    return Results.Ok(new
-    {
-        tenantId = subscription.TenantId,
-        companyName = subscription.CompanyName,
-        maxConcurrentUsers = subscription.MaxConcurrentUsers,
-        isActive = subscription.IsActive
-    });
-})
-.WithName("GetTenantConcurrentUserLimit")
-.WithOpenApi();
-
-app.MapGet("/catalogo/produtos", async (
-    ITenantContext tenantContext,
-    IProductCatalogRepository repository,
-    string? q,
-    int? limit,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.TenantId))
-    {
-        return Results.Unauthorized();
-    }
-
-    var effectiveLimit = limit.GetValueOrDefault(20);
-    if (effectiveLimit <= 0 || effectiveLimit > 100)
-    {
-        return Results.ValidationProblem(new Dictionary<string, string[]>
-        {
-            ["limit"] = ["limit must be between 1 and 100."]
-        });
-    }
-
-    var products = await repository.SearchProductsAsync(
-        tenantContext.TenantId,
-        q,
-        effectiveLimit,
-        cancellationToken);
-
-    return Results.Ok(products);
-})
-.WithName("SearchProducts")
-.WithOpenApi();
-
-app.MapGet("/catalogo/clientes", async (
-    ITenantContext tenantContext,
-    IClientCatalogRepository repository,
-    string? q,
-    int? limit,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.TenantId))
-    {
-        return Results.Unauthorized();
-    }
-
-    var effectiveLimit = limit.GetValueOrDefault(50);
-    if (effectiveLimit <= 0 || effectiveLimit > 200)
-    {
-        return Results.ValidationProblem(new Dictionary<string, string[]>
-        {
-            ["limit"] = ["limit must be between 1 and 200."]
-        });
-    }
-
-    var clients = await repository.SearchClientsAsync(
-        tenantContext.TenantId,
-        q,
-        effectiveLimit,
-        cancellationToken);
-
-    return Results.Ok(clients);
-})
-.WithName("SearchClients")
-.WithOpenApi();
-
-app.MapPost("/pedidos", async (
-    ITenantContext tenantContext,
-    CriarPedidoRequest request,
-    IMediator mediator,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.TenantId))
-    {
-        return Results.Unauthorized();
-    }
-
-    var errors = request.Validate();
-    if (errors.Count > 0)
-    {
-        return Results.ValidationProblem(errors);
-    }
-
-    var result = await mediator.Send(new CriarPedidoCommand(
-        tenantContext.TenantId,
-        request.ClienteId,
-        request.Itens,
-        request.CondicaoPagamento,
-        request.Observacao), cancellationToken);
-
-    return Results.Created($"/pedidos/{result.PedidoId}", new
-    {
-        pedidoId = result.PedidoId,
-        status = result.Status,
-        itensCount = result.ItensCount,
-        parcelasCount = result.ParcelasCount,
-        totalBruto = result.TotalBruto,
-        totalDesconto = result.TotalDesconto,
-        totalLiquido = result.TotalLiquido
-    });
-})
-.WithName("CreatePedido")
-.WithOpenApi();
-
-
-app.MapGet("/pedidos/{id}", async (
-    ITenantContext tenantContext,
-    Guid id,
-    PedidosDbContext db,
-    IPedidoCache pedidoCache,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.TenantId))
-    {
-        return Results.Unauthorized();
-    }
-
-    var pedido = await db.Pedidos
-        .Where(p => p.Id == id)
-        .Include(p => p.Itens)
-        .Include(p => p.Parcelas)
-        .Include(p => p.Status)
-        .FirstOrDefaultAsync(cancellationToken);
-
-    if (pedido is null)
-    {
-        // Fallback para ambiente de testes onde o DB InMemory pode ser diferente entre escopos.
-        if (!pedidoCache.TryGet(id, out var cached) || cached is null)
-        {
-            return Results.NotFound();
-        }
-        pedido = cached;
-    }
-
-    var itens = pedido.Itens.Select(i => new PedidoItemDto(
-        i.ProdutoId,
-        i.Sku,
-        i.Nome,
-        i.Quantidade,
-        i.PrecoUnitario,
-        i.Desconto,
-        i.Total)).ToList();
-
-    var parcelas = pedido.Parcelas.OrderBy(p => p.Numero).Select(p => new PedidoParcelaDto(
-        p.Numero,
-        p.DataVencimento,
-        p.Valor,
-        p.FormaPagamento)).ToList();
-
-    var response = new PedidoResponse(
-        pedido.Id,
-        pedido.TenantId,
-        pedido.ClienteId,
-        pedido.CriadoEm,
-        pedido.Status?.Codigo ?? "",
-        itens.Count,
-        parcelas.Count,
-        pedido.TotalBruto,
-        pedido.TotalDesconto,
-        pedido.TotalLiquido,
-        itens,
-        parcelas,
-        pedido.Observacao);
-
-    return Results.Ok(response);
-})
-.WithName("GetPedido")
-.WithOpenApi();
-
-// Endpoint para listar pedidos do tenant autenticado
-app.MapGet("/pedidos", async (
-    ITenantContext tenantContext,
-    PedidosDbContext db,
-    IPedidoCache pedidoCache,
-    string? clienteId,
-    string? status,
-    int? page,
-    int? pageSize,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.TenantId))
-    {
-        return Results.Unauthorized();
-    }
-
-    var query = db.Pedidos
-        .AsNoTracking()
-        .Include(p => p.Status)
-        .Include(p => p.Itens)
-        .Include(p => p.Parcelas)
-        .Where(p => p.TenantId == tenantContext.TenantId);
-
-    if (!string.IsNullOrWhiteSpace(clienteId))
-    {
-        query = query.Where(p => p.ClienteId == clienteId);
-    }
-    if (!string.IsNullOrWhiteSpace(status))
-    {
-        query = query.Where(p => p.Status != null && p.Status.Codigo == status);
-    }
-
-    int pageNumber = page.GetValueOrDefault(1);
-    int pageSizeNumber = pageSize.GetValueOrDefault(20);
-    if (pageNumber < 1) pageNumber = 1;
-    if (pageSizeNumber < 1 || pageSizeNumber > 100) pageSizeNumber = 20;
-
-    var pedidos = await query
-        .OrderByDescending(p => p.CriadoEm)
-        .Skip((pageNumber - 1) * pageSizeNumber)
-        .Take(pageSizeNumber)
-        .ToListAsync(cancellationToken);
-
-    if (pedidos.Count == 0)
-    {
-        // Fallback para ambiente de testes onde o DB InMemory pode variar por escopo.
-        var cachedQuery = pedidoCache.GetByTenant(tenantContext.TenantId)
-            .AsEnumerable();
-
-        if (!string.IsNullOrWhiteSpace(clienteId))
-        {
-            cachedQuery = cachedQuery.Where(p => p.ClienteId == clienteId);
-        }
-
-        if (!string.IsNullOrWhiteSpace(status))
-        {
-            cachedQuery = cachedQuery.Where(p => p.Status != null && p.Status.Codigo == status);
-        }
-
-        pedidos = cachedQuery
-            .OrderByDescending(p => p.CriadoEm)
-            .Skip((pageNumber - 1) * pageSizeNumber)
-            .Take(pageSizeNumber)
-            .ToList();
-    }
-
-    var result = pedidos.Select(p => new
-    {
-        pedidoId = p.Id,
-        tenantId = p.TenantId,
-        clienteId = p.ClienteId,
-        criadoEm = p.CriadoEm,
-        status = p.Status?.Codigo ?? string.Empty,
-        itensCount = p.Itens.Count,
-        parcelasCount = p.Parcelas.Count,
-        totalBruto = p.TotalBruto,
-        totalDesconto = p.TotalDesconto,
-        totalLiquido = p.TotalLiquido
-    });
-
-    return Results.Ok(result);
-})
-.WithName("ListPedidosQuery")
-.WithOpenApi();
-app.MapMethods("/auth/heartbeat", ["PATCH"], async (
-    ITenantContext tenantContext,
-    ISessionStore sessionStore,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.SessionId))
-    {
-        return Results.Unauthorized();
-    }
-
-    await sessionStore.HeartbeatAsync(tenantContext.SessionId, tenantContext.TenantId!, cancellationToken);
-    return Results.Ok(new { message = "Session renewed.", sessionId = tenantContext.SessionId });
-})
-.WithName("SessionHeartbeat")
-.WithOpenApi();
-
-app.MapGet("/admin/sessions", async (
-    ITenantContext tenantContext,
-    ISessionStore sessionStore,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant)
-    {
-        return Results.Unauthorized();
-    }
-
-    var sessions = await sessionStore.GetActiveSessionsAsync(tenantContext.TenantId!, cancellationToken);
-    return Results.Ok(sessions.Select(s => new
-    {
-        sessionId = s.SessionId,
-        userId = s.UserId,
-        tenantId = s.TenantId,
-        loginAt = s.LoginAt,
-        lastHeartbeatAt = s.LastHeartbeatAt
-    }));
-})
-.WithName("GetActiveSessions")
-.WithOpenApi();
-
-app.MapPost("/admin/sessions/evict", async (
-    ITenantContext tenantContext,
-    ISessionStore sessionStore,
-    IRefreshTokenStore refreshTokenStore,
-    ISessionAuditEventRepository auditRepo,
-    EvictRequest request,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant)
-    {
-        return Results.Unauthorized();
-    }
-
-    if (string.IsNullOrWhiteSpace(request?.SessionId))
-    {
-        return Results.ValidationProblem(new Dictionary<string, string[]>
-        {
-            ["sessionId"] = new[] { "sessionId is required" }
-        });
-    }
-
-    // Remove session from store
-    await sessionStore.RemoveAsync(request.SessionId, tenantContext.TenantId!, cancellationToken);
-
-    // Revoke all refresh tokens for the evicted user and optionally the one provided
-    if (!string.IsNullOrWhiteSpace(request.UserId))
-    {
-        refreshTokenStore.RevokeAllForUser(request.UserId);
-    }
-    else if (!string.IsNullOrWhiteSpace(request.RefreshToken))
-    {
-        refreshTokenStore.Revoke(request.RefreshToken);
-    }
-
-    // Audit eviction event
-    var evictAudit = new Versatus.ForcaVendas.Domain.Auditoria.SessionAuditEvent(
-        Id: Guid.NewGuid().ToString(),
-        UserId: request.UserId ?? tenantContext.UserId ?? string.Empty,
-        TenantId: tenantContext.TenantId!,
-        EventType: Versatus.ForcaVendas.Domain.Auditoria.SessionAuditEventType.Eviction,
-        Timestamp: DateTimeOffset.UtcNow
-    );
-    await auditRepo.AddAsync(evictAudit, cancellationToken);
-
-    return Results.Ok(new { message = "Session evicted", sessionId = request.SessionId });
-})
-    .WithName("EvictSession")
-    .WithOpenApi();
-
-app.MapPost("/auth/logout", async (
-    ITenantContext tenantContext,
-    IRefreshTokenStore refreshTokenStore,
-    ISessionStore sessionStore,
-    ISessionAuditEventRepository auditRepo,
-    HttpContext httpContext,
-    LogoutRequest? request,
-    CancellationToken cancellationToken) =>
-{
-    if (!tenantContext.HasTenant || string.IsNullOrWhiteSpace(tenantContext.SessionId))
-    {
-        return Results.Unauthorized();
-    }
-
-    if (request is not null && !string.IsNullOrWhiteSpace(request.RefreshToken))
-    {
-        refreshTokenStore.Revoke(request.RefreshToken);
-    }
-
-    await sessionStore.RemoveAsync(tenantContext.SessionId, tenantContext.TenantId!, cancellationToken);
-    // Audit logout event
-    var auditEvent = new Versatus.ForcaVendas.Domain.Auditoria.SessionAuditEvent(
-        Id: Guid.NewGuid().ToString(),
-        UserId: tenantContext.UserId ?? string.Empty,
-        TenantId: tenantContext.TenantId!,
-        EventType: Versatus.ForcaVendas.Domain.Auditoria.SessionAuditEventType.Logout,
-        Timestamp: DateTimeOffset.UtcNow,
-        IpAddress: httpContext.Connection.RemoteIpAddress?.ToString(),
-        UserAgent: httpContext.Request.Headers["User-Agent"].ToString()
-    );
-    await auditRepo.AddAsync(auditEvent, cancellationToken);
-    return Results.Ok(new { message = "Logged out", sessionId = tenantContext.SessionId });
-})
-.WithName("Logout")
-.WithOpenApi();
+app.MapAuthSessionEndpoints();
+app.MapCatalogoEndpoints();
+app.MapPedidosEndpoints();
 
 app.MapControllers();
 app.Run();
